refactor(processes): extract the low-level Win32 kill helpers to their own module (Closes #773)
