+++ src/main/processes/kill.ts
@@ -1,7 +1,5 @@
-import { execFile, type ChildProcess } from 'child_process'
 import path from 'path'
 
-import { writeAppErrorLog } from '../errorLog'
 import {
   getActiveStoredProfile,
   getProfileTrackablePaths,
@@ -9,13 +7,7 @@ import {
   isUtilityEnabled
 } from '../profiles'
 import { getStoredStringRecord } from '../store'
-import {
-  getErrorMessage,
-  getExeName,
-  isValidExePath,
-  normalizePathForComparison,
-  pathsEqual
-} from '../utils'
+import { getExeName, isValidExePath, normalizePathForComparison, pathsEqual } from '../utils'
 
 import {
   abortActiveLaunches,
@@ -29,33 +21,20 @@ import {
 } from './state'
 import { publishRunningApps } from './running'
 import { invalidateProcessNameCache, readRunningProcessNames } from './tasklist'
-import type { KillFailure, KillFailureReason, KillProfileAppsOptions, KillResult } from './types'
-
-// Generous for a healthy system (the query usually returns in tens of ms) but
-// bounded so a wedged WMI service cannot hang a kill request forever.
-const WMI_LOOKUP_TIMEOUT_MS = 3000
-
-export interface KillAttemptResult {
-  processName: string
-  success: boolean
-  appPath?: string
-  gameKey?: string
-  error?: string
-  accessDenied?: boolean
-  notFound?: boolean
-  staleTask?: boolean
-  stillRunning?: boolean
-  /**
-   * Set only when this attempt positively observed a process under
-   * `processName`: PIDs discovered at its path, a tracked child that was really
-   * there, or an `/IM` taskkill that found something to act on.
-   *
-   * Deliberately NOT the inverse of `notFound`. A WMI lookup that errors out
-   * returns neither, and treating that absence as evidence would let a failed
-   * lookup vouch for a sibling it never looked at (CodeRabbit on #818).
-   */
-  targetConfirmed?: boolean
-}
+import type {
+  KillAttemptResult,
+  KillFailure,
+  KillFailureReason,
+  KillProfileAppsOptions,
+  KillResult
+} from './types'
+import {
+  findProcessIdsByExecutablePath,
+  isFullExePath,
+  isPathScopedExe,
+  killProcessByImageName,
+  killProcessTree
+} from './win32KillUtils'
 
 // Hardcoded list of companion process names for utilities that spawn background
 // agents under a name that differs from (or is not derived from) their
@@ -64,287 +43,6 @@ export interface KillAttemptResult {
 // need an entry here.
 const UTILITY_COMPANION_PROCESS_NAMES: Record<string, string[]> = {
   garage61: ['Garage61 telemetry agent.exe']
-}
-
-function isAccessDeniedMessage(message: string) {
-  return /(access is denied|permission denied|administrator|elevat)/i.test(message)
-}
-
-function isNotFoundMessage(message: string) {
-  return /not found|no running instance/i.test(message)
-}
-
-function isStaleTaskMessage(message: string) {
-  return /no running instance/i.test(message)
-}
-
-function runTaskkill(args: string[], description: string) {
-  return new Promise<{
-    success: boolean
-    detail?: string
-    accessDenied?: boolean
-    notFound?: boolean
-    staleTask?: boolean
-  }>((resolve) => {
-    execFile('taskkill', args, { windowsHide: true }, (error, stdout, stderr) => {
-      if (!error) {
-        resolve({ success: true })
-        return
-      }
-
-      const detail = stderr.trim() || stdout.trim() || error.message
-      const notFound = isNotFoundMessage(detail)
-      const staleTask = isStaleTaskMessage(detail)
-      const accessDenied = isAccessDeniedMessage(detail)
-
-      if (!notFound) {
-        console.error(`Failed to ${description}: ${detail}`)
-        writeAppErrorLog('kill', `Failed to ${description}: ${detail}`)
-      }
-
-      resolve({
-        success: notFound,
-        detail,
-        accessDenied,
-        notFound,
-        staleTask
-      })
-    })
-  })
-}
-
-/**
- * Guard that distinguishes a fully-qualified exe path (e.g.
- * `C:\Tools\app.exe`) from a bare process name (e.g. `app.exe`). Only full
- * paths are eligible for path-scoped kills via WMI `ExecutablePath` matching —
- * bare names fall back to the less precise `/IM` image-name kill to avoid
- * refusing to kill a process whose path we cannot verify.
- */
-function isFullExePath(appPath: string | undefined): appPath is string {
-  return (
-    typeof appPath === 'string' && path.basename(appPath) !== appPath && isValidExePath(appPath)
-  )
-}
-
-/**
- * True when `appPath` is a directory-qualified exe path (has a parent dir),
- * judged by SHAPE ALONE. Unlike {@link isFullExePath} it does NOT stat the
- * filesystem. Use this to decide whether basename fallback cleanup is allowed:
- * scoping must not depend on the exe still being present, or a full-path game
- * whose exe was removed or is momentarily inaccessible mid-close would fall back
- * to matching by image name and delete a DIFFERENT game's same-named companion
- * tracked at another path (#677).
- */
-function isPathScopedExe(appPath: string | undefined): appPath is string {
-  return typeof appPath === 'string' && path.basename(appPath) !== appPath
-}
-
-function parseProcessIds(output: string) {
-  const trimmedOutput = output.trim()
-
-  if (!trimmedOutput) {
-    return []
-  }
-
-  try {
-    const parsed = JSON.parse(trimmedOutput) as unknown
-    const values = Array.isArray(parsed) ? parsed : [parsed]
-
-    return values
-      .map((value) => (typeof value === 'number' ? value : Number(value)))
-      .filter((value) => Number.isSafeInteger(value) && value > 0)
-  } catch {
-    return trimmedOutput
-      .split(/\r?\n/)
-      .map((line) => Number(line.trim()))
-      .filter((value) => Number.isSafeInteger(value) && value > 0)
-  }
-}
-
-function findProcessIdsByExecutablePath(processName: string, appPath: string) {
-  return new Promise<{
-    processIds: number[]
-    detail?: string
-    accessDenied?: boolean
-  }>((resolve) => {
-    const script = [
-      // Both the target path and the process name are injected via environment
-      // variables rather than interpolated into the script string. This prevents
-      // a value containing single-quotes or PowerShell metacharacters from
-      // breaking out of a string literal or injecting arbitrary commands.
-      '$target = $env:SIMLAUNCHER_TARGET_PROCESS_PATH',
-      '$name = $env:SIMLAUNCHER_TARGET_PROCESS_NAME',
-      '$targetPath = [System.IO.Path]::GetFullPath($target)',
-      // Match the process name in PowerShell with -ieq rather than in a WQL
-      // `Name = '...'` filter. WQL string-literal quote escaping is ambiguous and
-      // version-dependent (SQL-style doubling vs backslash), and getting it wrong
-      // silently breaks the lookup for exe names containing a single quote — the
-      // exact case this guards (#531). Comparing $_.Name to the env-injected $name
-      // in the host language sidesteps WQL escaping entirely and handles any
-      // character. The (rare, user-initiated) full-process enumeration is bounded
-      // by WMI_LOOKUP_TIMEOUT_MS; precision still comes from the ExecutablePath match.
-      'Get-CimInstance Win32_Process |',
-      '  Where-Object { $_.Name -ieq $name -and $_.ExecutablePath -and ([System.IO.Path]::GetFullPath($_.ExecutablePath) -ieq $targetPath) } |',
-      '  Select-Object -ExpandProperty ProcessId |',
-      '  ConvertTo-Json -Compress'
-    ].join('\n')
-
-    execFile(
-      'powershell.exe',
-      ['-NoProfile', '-NonInteractive', '-ExecutionPolicy', 'Bypass', '-Command', script],
-      {
-        windowsHide: true,
-        // A hung or slow WMI query (slow disk, process-heavy system) must not
-        // stall the kill pipeline indefinitely. On timeout this surfaces as a
-        // clean kill failure — deliberately NOT a `taskkill /IM` fallback,
-        // which would break the path-scoping safety guarantee and kill
-        // same-named processes the user started outside SimLauncher (#503).
-        timeout: WMI_LOOKUP_TIMEOUT_MS,
-        env: {
-          ...process.env,
-          SIMLAUNCHER_TARGET_PROCESS_PATH: path.resolve(appPath),
-          SIMLAUNCHER_TARGET_PROCESS_NAME: processName
-        }
-      },
-      (error, stdout, stderr) => {
-        if (error) {
-          // execFile sets `killed` when it terminated the child itself —
-          // with a plain timeout option that means the deadline elapsed.
-          const detail = error.killed
-            ? `Process lookup timed out after ${WMI_LOOKUP_TIMEOUT_MS / 1000} seconds.`
-            : stderr.trim() || stdout.trim() || error.message
-          console.error(`Failed to find process IDs for ${appPath}: ${detail}`)
-          writeAppErrorLog('kill', `Failed to find process IDs for ${appPath}: ${detail}`)
-          resolve({ processIds: [], detail, accessDenied: isAccessDeniedMessage(detail) })
-          return
-        }
-
-        resolve({ processIds: parseProcessIds(stdout) })
-      }
-    )
-  })
-}
-
-async function killProcessTree(
-  child: ChildProcess,
-  appPath: string,
-  gameKey?: string
-): Promise<KillAttemptResult> {
-  const processName = getExeName(appPath)
-
-  if (process.platform === 'win32' && child.pid) {
-    const result = await runTaskkill(
-      ['/PID', String(child.pid), '/T', '/F'],
-      `kill process tree for ${appPath}`
-    )
-    return {
-      processName,
-      appPath,
-      gameKey,
-      success: result.success,
-      error: result.detail,
-      accessDenied: result.accessDenied,
-      notFound: result.notFound,
-      staleTask: result.staleTask,
-      // A tracked child we hold a PID for was a real target unless taskkill
-      // reports it had already exited.
-      targetConfirmed: !result.notFound
-    }
-  }
-
-  try {
-    child.kill()
-    return { processName, appPath, gameKey, success: true, targetConfirmed: true }
-  } catch (err) {
-    const error = getErrorMessage(err)
-    console.error(`Error killing ${appPath}:`, err)
-    writeAppErrorLog('kill', `Error killing ${appPath}: ${error}`)
-    return {
-      processName,
-      appPath,
-      gameKey,
-      success: false,
-      error,
-      accessDenied: isAccessDeniedMessage(error)
-    }
-  }
-}
-
-async function killProcessByImageName(
-  processName: string,
-  appPath?: string,
-  gameKey?: string
-): Promise<KillAttemptResult> {
-  if (process.platform !== 'win32') {
-    return { processName, appPath, gameKey, success: true }
-  }
-
-  if (isFullExePath(appPath)) {
-    const targetAppPath = appPath
-    const { processIds, detail, accessDenied } = await findProcessIdsByExecutablePath(
-      processName,
-      targetAppPath
-    )
-
-    if (detail) {
-      return {
-        processName,
-        appPath: targetAppPath,
-        gameKey,
-        success: false,
-        error: detail,
-        accessDenied
-      }
-    }
-
-    if (processIds.length === 0) {
-      // Elevated processes can expose null ExecutablePath in WMI, so they are silently
-      // filtered out by the Where-Object clause; treat this as notFound rather than error.
-      return { processName, appPath: targetAppPath, gameKey, success: true, notFound: true }
-    }
-
-    const results = await Promise.all(
-      processIds.map((processId) =>
-        runTaskkill(
-          ['/PID', String(processId), '/T', '/F'],
-          `kill companion process ${targetAppPath}`
-        )
-      )
-    )
-    const failedResult = results.find((result) => !result.success && !result.notFound)
-
-    return {
-      processName,
-      appPath: targetAppPath,
-      gameKey,
-      success: !failedResult,
-      error: failedResult?.detail,
-      accessDenied: failedResult?.accessDenied,
-      notFound: results.every((result) => result.notFound),
-      staleTask: results.every((result) => result.staleTask),
-      // PIDs were discovered at this exact path, so a process under this image
-      // name demonstrably existed.
-      targetConfirmed: true
-    }
-  }
-
-  const result = await runTaskkill(
-    ['/IM', processName, '/T', '/F'],
-    `kill companion process ${processName}`
-  )
-  return {
-    processName,
-    appPath,
-    gameKey,
-    success: result.success,
-    error: result.detail,
-    accessDenied: result.accessDenied,
-    notFound: result.notFound,
-    staleTask: result.staleTask,
-    // taskkill either killed something or was denied by something. Any other
-    // failure is ambiguous, so it does not count as evidence.
-    targetConfirmed: result.success || result.accessDenied === true
-  }
 }
 
 function clearUnclosedProcess(
